chatbi-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.36
pymysql>=1.1.1
aiomysql>=0.2.0
neo4j>=5.26.0
psycopg2-binary>=2.9.10
pgvector>=0.3.5
# LangChain Ecosystem (Modern Split Architecture)
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
openai>=1.54.0
dashscope>=1.20.0
requests>=2.32.0
httpx>=0.27.0
mcp>=1.0.0
langgraph>=0.2.0
websockets>=12.0
tqdm>=4.66.0
